+++ tomcat/src/main/java/nextstep/jwp/presenstation/LoginHandler.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.presenstation;
 
 import java.util.Map;
 import java.util.Optional;
@@ -6,10 +6,11 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.ResponseEntity;
-import org.apache.coyote.http11.message.HttpHeaders;
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpStatus;
-import org.apache.coyote.http11.message.RequestBody;
+import org.apache.coyote.http11.handler.AbstractRequestHandler;
+import org.apache.coyote.http11.http.HttpHeaders;
+import org.apache.coyote.http11.http.HttpRequest;
+import org.apache.coyote.http11.http.HttpStatus;
+import org.apache.coyote.http11.http.RequestBody;
 import org.apache.coyote.http11.session.SessionManager;
 import org.apache.coyote.util.FileUtil;
