+++ tomcat/src/main/java/nextstep/jwp/presenstation/RegisterHandler.java
@@ -1,15 +1,16 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.presenstation;
 
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.ResponseEntity;
-import org.apache.coyote.http11.message.HttpHeaders;
-import org.apache.coyote.http11.message.HttpMethod;
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpStatus;
-import org.apache.coyote.http11.message.RequestBody;
+import org.apache.coyote.http11.handler.AbstractRequestHandler;
+import org.apache.coyote.http11.http.HttpHeaders;
+import org.apache.coyote.http11.http.HttpMethod;
+import org.apache.coyote.http11.http.HttpRequest;
+import org.apache.coyote.http11.http.HttpStatus;
+import org.apache.coyote.http11.http.RequestBody;
 import org.apache.coyote.util.FileUtil;
 
 public class RegisterHandler extends AbstractRequestHandler {
