+++ tomcat/src/main/java/org/apache/coyote/http11/handler/AbstractRequestHandler.java
@@ -1,9 +1,10 @@
 package org.apache.coyote.http11.handler;
 
+import nextstep.jwp.presenstation.RequestHandler;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.ResponseEntity;
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpStatus;
+import org.apache.coyote.http11.http.HttpRequest;
+import org.apache.coyote.http11.http.HttpStatus;
 import org.apache.coyote.util.FileUtil;
 
 public abstract class AbstractRequestHandler implements RequestHandler {
