+++ tomcat/src/main/java/org/apache/coyote/http11/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.message;
+package org.apache.coyote.http11.http;
 
 public enum HttpMethod {
     GET,
